Get the game back to a working state
It will check if the data files are already in place and try to download it if not available.
It is not, however, handling download and exrtaction failures. If you have a partiatlly extracted Doom.WAD, the game will remain in an inconsistent state and your only option is to re-install it.
Similarly, if the download fails, you won't be notified and will have to close the app and open again.